refactor: local UserMenuLogin to manage Login button
Related to https://github.com/CartoDB/carto-react-lib/pull/44
